+++ packages/sdk/src/configs/token-list.ts
@@ -30,8 +30,8 @@ const DEFAULT_TOKEN_LIST: TokenListConfig = {
   [ChainType.ACALA]: {
     ACA: createTokenConfig('ACA', 'ACA', 12, new FixedPointNumber(0.1, 12)),
     AUSD: createTokenConfig('AUSD', 'aUSD', 12, new FixedPointNumber(0.1, 12)),
-    DOT: createTokenConfig('DOT', 'DOT', 12, new FixedPointNumber(0.01, 12)),
-    LDOT: createTokenConfig('LDOT', 'LDOT', 12, new FixedPointNumber(0.05, 12))
+    DOT: createTokenConfig('DOT', 'DOT', 10, new FixedPointNumber(0.01, 12)),
+    LDOT: createTokenConfig('LDOT', 'LDOT', 10, new FixedPointNumber(0.05, 12))
   },
   [ChainType.KARURA]: {
     KAR: createTokenConfig('KAR', 'KAR', 12, new FixedPointNumber(0.1, 12)),
@@ -40,13 +40,16 @@ const DEFAULT_TOKEN_LIST: TokenListConfig = {
     LKSM: createTokenConfig('LKSM', 'LKSM', 12, new FixedPointNumber(50 * 0.00001, 12)),
     BNC: createTokenConfig('BNC', 'BNC', 12, new FixedPointNumber(800 * 0.00001, 12)),
     VSKSM: createTokenConfig('VSKSM', 'vsKSM', 12, new FixedPointNumber(10 * 0.00001, 12)),
-    PHA: createTokenConfig('PHA', 'PHA', 12, new FixedPointNumber(4000 * 0.00001, 12))
+    PHA: createTokenConfig('PHA', 'PHA', 12, new FixedPointNumber(4000 * 0.00001, 12)),
+    KINT: createTokenConfig('KINT', 'KINT', 12, new FixedPointNumber(13333 * 0.00001, 12)),
+    KBTC: createTokenConfig('KBTC', 'KBTC', 12, new FixedPointNumber(66 * 0.00001, 8)),
+    TAI: createTokenConfig('TAI', 'TAI', 12, new FixedPointNumber(1, 12))
   },
   [ChainType.MANDALA]: {
     ACA: createTokenConfig('ACA', 'ACA', 12, new FixedPointNumber(0.1, 12)),
     AUSD: createTokenConfig('AUSD', 'aUSD', 12, new FixedPointNumber(0.1, 12)),
-    DOT: createTokenConfig('DOT', 'DOT', 12, new FixedPointNumber(0.01, 12)),
-    LDOT: createTokenConfig('LDOT', 'LDOT', 12, new FixedPointNumber(0.05, 12)),
+    DOT: createTokenConfig('DOT', 'DOT', 10, new FixedPointNumber(0.01, 12)),
+    LDOT: createTokenConfig('LDOT', 'LDOT', 10, new FixedPointNumber(0.05, 12)),
     KAR: createTokenConfig('KAR', 'KAR', 12, new FixedPointNumber(0.1, 12)),
     KUSD: createTokenConfig('KUSD', 'kUSD', 12, new FixedPointNumber(0.01, 12)),
     KSM: createTokenConfig('KSM', 'KSM', 12, new FixedPointNumber(10 * 0.00001, 12)),