+++ packages/sdk/src/wallet/price-provider/default-token-price-fetch-source-config.ts
@@ -4,7 +4,8 @@ import { PriceProviderType } from './types';
 
 export const defaultTokenPriceFetchSource: TokenPriceFetchSource = {
   [ChainType.ACALA]: {
-    DOT: PriceProviderType.MARKET
+    DOT: PriceProviderType.MARKET,
+    'lc://13': PriceProviderType.ORACLE
   },
   [ChainType.KARURA]: {
     KAR: PriceProviderType.MARKET,
@@ -14,6 +15,7 @@ export const defaultTokenPriceFetchSource: TokenPriceFetchSource = {
   [ChainType.MANDALA]: {
     DOT: PriceProviderType.MARKET,
     KAR: PriceProviderType.MARKET,
+    KSM: PriceProviderType.MARKET,
     BNC: PriceProviderType.MARKET
   }
 };