+++ packages/sdk/src/wallet/price-provider/types.ts
@@ -1,12 +1,13 @@
 import { Observable } from 'rxjs';
-import { FixedPointNumber as FN } from '@acala-network/sdk-core';
+import { FixedPointNumber as FN, Token } from '@acala-network/sdk-core';
 
 export interface PriceProvider {
-  subscribe(currency: string): Observable<FN>;
-  query(currency: string): Promise<FN>;
+  subscribe(currency: Token): Observable<FN>;
+  query(currency: Token): Promise<FN>;
 }
 
 export enum PriceProviderType {
   'MARKET', // query price form market
-  'ORACLE' // query oracle feed prices
+  'ORACLE', // query oracle feed prices
+  'DEX' // query price form dex
 }