+++ packages/type-definitions/src/primitives.ts
@@ -26,7 +26,10 @@ export default {
         // 150: Reserved for CASH
         // 168 - 255: Kusama parachain tokens
         BNC: 168,
-        VSKSM: 169
+        VSKSM: 169,
+        PHA: 170,
+        KINT: 171,
+        KBTC: 172
       }
     },
     DexShare: {
@@ -63,7 +66,8 @@ export default {
       symbol: 'Vec<u8>',
       decimals: 'u8',
       minimalBalance: 'Balance'
-    }
+    },
+    NumberOrHex: 'u128'
   },
   typesAlias: {
     rewards: {